fix: 10개로 최근 본 상품 안가져오는 문제 해결

+++ app/src/main/java/woowacourse/shopping/data/dao/RecentProductDao.kt
@@ -9,21 +9,29 @@ import woowacourse.shopping.data.entity.RecentProductWithProduct
 
 @Dao
 interface RecentProductDao {
-    @Insert
+    @Insert(onConflict = androidx.room.OnConflictStrategy.REPLACE)
     fun insert(recentProductEntity: RecentProductEntity)
 
     @Query(
         "DELETE FROM recent_product " +
-            "WHERE viewTime = (SELECT MAX(viewTime) FROM recent_product LIMIT 1)",
+            "WHERE viewTime = (SELECT MIN(viewTime) FROM recent_product LIMIT 1)",
     )
     fun deleteLast()
 
-    @Query("SELECT * FROM recent_product")
+    @Query(
+        "SELECT * FROM recent_product " +
+            "INNER JOIN product ON recent_product.product_id = product.id",
+    )
     fun findAll(): List<RecentProductWithProduct>
 
     @Transaction
     fun save(recentProductEntity: RecentProductEntity) {
-        deleteLast()
+        if (!doesProductExist(recentProductEntity.productId)) {
+            deleteLast()
+        }
         insert(recentProductEntity)
     }
+
+    @Query("SELECT EXISTS(SELECT 1 FROM recent_product WHERE product_id = :productId LIMIT 1)")
+    fun doesProductExist(productId: Long): Boolean
 }
